+++ packages/event-handler/src/index.ts
@@ -4,3 +4,4 @@ export * from "./chatHandler";
 export * from "./activity";
 export * from "./server";
 export * from "./remote";
+export * from "./tools"