+++ packages/activity-manager/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sugarch/bc-activity-manager
 
+## 0.4.6
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @sugarch/bc-event-handler@0.6.2
+
 ## 0.4.5
 
 ### Patch Changes